--- patternList.txt
@@ -1,23 +1,0 @@
-walkingOnes
-walkingZeros
-walkingOnesZeros
-walkingOnes456
-walkingZeros456
-walkingOnesZeros456
-cornerSharing
-edges
-intraEdges
-testCard6
-flooding
-allTowers
-zeroMax
-regionSumPins0
-regionSumPins1
-regionSumPins2
-regionSumPins3
-test
-zeros
-count
-testCrateNumber
-random
-tplf
