+++ plugins/BuildFile.xml
@@ -1,0 +1,14 @@
+<use name="FWCore/Framework"/>
+<use name="FWCore/PluginManager"/>
+<use name="FWCore/ParameterSet"/>
+<use   name="DataFormats/L1Trigger"/>
+<use   name="CondTools/L1Trigger"/>
+<use   name="CondTools/L1Trigger"/>
+<use   name="DataFormats/L1CaloTrigger"/>
+<use   name="DataFormats/EcalDigi"/>
+<use   name="DataFormats/EcalDetId"/>
+<use   name="DataFormats/HcalDigi"/>
+<use   name="DataFormats/HcalDetId"/>
+<use   name="CondFormats/L1TObjects"/>
+<use   name="L1Trigger/RegionalCaloTrigger">
+<flags EDM_PLUGIN="1"/>
