--- rootlogon.C
@@ -1,4 +1,0 @@
-{
-//TFile *file0 = new TFile("etaPhiSpaces.root"); 
-G__loadfile("plotFilter.C");
-}
