--- ttbarEventList.txt
@@ -1,36 +1,0 @@
-   
-#     '/store/relval/CMSSW_2_1_8/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/IDEAL_V9_v1/0002/1A2B8C4B-9082-DD11-A33F-000423D98EC8.root', '/store/relval/CMSSW_2_1_8/RelValTTbar/ALCARECO/IDEAL_V9_StreamALCARECOEcalCalElectron_v1/0003/967A6453-0283-DD11-9447-000423D98930.root',
-#                                       '/store/relval/CMSSW_2_1_9/RelValTTbar/ALCARECO/STARTUP_V7_StreamALCARECOHcalCalGammaJet_v2/0003/58B63CB7-8688-DD11-B581-001617DC1F70.root',
-#'/store/relval/CMSSW_2_1_8/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/IDEAL_V9_v1/0002/04983078-9082-DD11-BB8C-0019DB2F3F9B.root',
-#       '/store/relval/CMSSW_2_1_8/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/IDEAL_V9_v1/0002/22F69384-9082-DD11-8C09-000423D98750.root',
-#   '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/102167CE-FA7D-DD11-B1AD-000423D6B42C.root',
-#       '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/103905CD-FA7D-DD11-8347-000423D9989E.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/14A717C6-FA7D-DD11-9645-000423D99896.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/1A2DFB65-FA7D-DD11-BE81-001617E30D0A.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/343B6E98-FA7D-DD11-AA5B-000423D9890C.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/36A3F4CF-FA7D-DD11-B918-000423D98F98.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/3A1FB369-FA7D-DD11-81DC-000423D6BA18.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/3AF210A2-FA7D-DD11-9CAF-000423D99658.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/4080F7A4-FA7D-DD11-8A7C-000423D94A04.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/40B0DD75-FA7D-DD11-81AD-000423D98844.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/44064AA5-FA7D-DD11-AD1D-000423D99EEE.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/4C7C3DA2-FA7D-DD11-B98A-000423D951D4.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/4CD4FAA5-FA7D-DD11-86E9-000423D99CEE.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/56856F71-FA7D-DD11-A5AE-000423D992A4.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/56FC3F75-FA7D-DD11-9927-000423D94C68.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/6A0D8295-FA7D-DD11-93B5-000423D99AAE.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/70CD5173-FA7D-DD11-BAB0-000423D95220.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/748F20C5-FA7D-DD11-BB62-000423D98844.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/7C9A20A3-FA7D-DD11-90BB-000423D992A4.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/8A9EFA9D-FA7D-DD11-8105-000423D98834.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/9E511675-FA7D-DD11-B851-000423D985B0.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/A4B75BA0-FA7D-DD11-9126-000423D99160.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/A6B7C5FC-FA7D-DD11-8A5B-000423D6B5C4.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/A6BE69FC-FA7D-DD11-A628-000423D98AF0.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/B41E146C-FA7D-DD11-B1FC-000423D98B5C.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/CE2F51CC-FA7D-DD11-83FF-000423D944DC.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/E6312B74-FA7D-DD11-A6DF-000423DD2F34.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/F83748C6-FA7D-DD11-B6C4-000423D998BA.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/FC12F873-FA7D-DD11-B9D9-000423D98DC4.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0002/86BC4829-437E-DD11-9BC1-001617C3B77C.root',
-#        '/store/relval/CMSSW_2_1_7/RelValTTbar/GEN-SIM-DIGI-RAW-HLTDEBUG-RECO/STARTUP_V7_v1/0001/F4E80FCF-FA7D-DD11-A0E2-000423D94C68.root'
