--- walkingList.txt
@@ -1,6 +1,0 @@
-walkingOnes
-walkingZeros
-walkingOnesZeros
-walkingOnes456
-walkingZeros456
-walkingOnesZeros456
